Migrate /test/apm_api_integration/tests/dependencies to be deployment-agnostic API tests
