+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/dependencies/top_dependencies.spec.ts
@@ -7,16 +7,16 @@
 import expect from '@kbn/expect';
 import { APIReturnType } from '@kbn/apm-plugin/public/services/rest/create_call_apm_api';
 import { NodeType, DependencyNode } from '@kbn/apm-plugin/common/connections';
-import { FtrProviderContext } from '../../common/ftr_provider_context';
+import { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
+import type { DeploymentAgnosticFtrProviderContext } from '../../../../ftr_provider_context';
 import { dataConfig, generateData } from './generate_data';
-import { roundNumber } from '../../utils';
+import { roundNumber } from '../utils/common';
 
 type TopDependencies = APIReturnType<'GET /internal/apm/dependencies/top_dependencies'>;
 
-export default function ApiTest({ getService }: FtrProviderContext) {
-  const registry = getService('registry');
-  const apmApiClient = getService('apmApiClient');
-  const apmSynthtraceEsClient = getService('apmSynthtraceEsClient');
+export default function ApiTest({ getService }: DeploymentAgnosticFtrProviderContext) {
+  const apmApiClient = getService('apmApi');
+  const synthtrace = getService('synthtrace');
 
   const start = new Date('2021-01-01T00:00:00.000Z').getTime();
   const end = new Date('2021-01-01T00:15:00.000Z').getTime() - 1;
@@ -37,24 +37,21 @@ export default function ApiTest({ getService }: FtrProviderContext) {
     });
   }
 
-  registry.when(
-    'Top dependencies when data is not loaded',
-    { config: 'basic', archives: [] },
-    () => {
+  describe('Top dependencies', () => {
+    describe('when data is not loaded', () => {
       it('handles empty state', async () => {
         const { status, body } = await callApi();
         expect(status).to.be(200);
         expect(body.dependencies).to.empty();
       });
-    }
-  );
+    });
 
-  // FLAKY: https://github.com/elastic/kibana/issues/177126
-  registry.when('Top dependencies', { config: 'basic', archives: [] }, () => {
     describe('when data is generated', () => {
       let topDependencies: TopDependencies;
+      let apmSynthtraceEsClient: ApmSynthtraceEsClient;
 
       before(async () => {
+        apmSynthtraceEsClient = await synthtrace.createApmSynthtraceEsClient();
         await generateData({ apmSynthtraceEsClient, start, end });
         const response = await callApi();
         topDependencies = response.body;
