+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/index.ts
@@ -13,5 +13,6 @@ export default function apmApiIntegrationTests({
   describe('APM', function () {
     loadTestFile(require.resolve('./agent_explorer'));
     loadTestFile(require.resolve('./custom_dashboards'));
+    loadTestFile(require.resolve('./dependencies'));
   });
 }
